refactor(app): move stripe under app pkg
