+++ cmd/server/main.go
@@ -20,9 +20,9 @@ import (
 	apppkg "github.com/openmeterio/openmeter/openmeter/app"
 	appadapter "github.com/openmeterio/openmeter/openmeter/app/adapter"
 	appservice "github.com/openmeterio/openmeter/openmeter/app/service"
-	"github.com/openmeterio/openmeter/openmeter/appstripe"
-	appstripeadapter "github.com/openmeterio/openmeter/openmeter/appstripe/adapter"
-	appstripeservice "github.com/openmeterio/openmeter/openmeter/appstripe/service"
+	appstripe "github.com/openmeterio/openmeter/openmeter/app/stripe"
+	appstripeadapter "github.com/openmeterio/openmeter/openmeter/app/stripe/adapter"
+	appstripeservice "github.com/openmeterio/openmeter/openmeter/app/stripe/service"
 	"github.com/openmeterio/openmeter/openmeter/customer"
 	customeradapter "github.com/openmeterio/openmeter/openmeter/customer/adapter"
 	customerservice "github.com/openmeterio/openmeter/openmeter/customer/service"
