+++ openmeter/app/stripe/adapter/adapter.go
@@ -8,10 +8,10 @@ import (
 
 	"github.com/openmeterio/openmeter/openmeter/app"
 	appentity "github.com/openmeterio/openmeter/openmeter/app/entity"
-	"github.com/openmeterio/openmeter/openmeter/appstripe"
-	stripeclient "github.com/openmeterio/openmeter/openmeter/appstripe/client"
-	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
-	appstripeobserver "github.com/openmeterio/openmeter/openmeter/appstripe/observer"
+	appstripe "github.com/openmeterio/openmeter/openmeter/app/stripe"
+	stripeclient "github.com/openmeterio/openmeter/openmeter/app/stripe/client"
+	appstripeentity "github.com/openmeterio/openmeter/openmeter/app/stripe/entity"
+	appstripeobserver "github.com/openmeterio/openmeter/openmeter/app/stripe/observer"
 	"github.com/openmeterio/openmeter/openmeter/customer"
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/secret"
