+++ openmeter/app/stripe/observer/customer.go
@@ -9,8 +9,8 @@ import (
 	appentity "github.com/openmeterio/openmeter/openmeter/app/entity"
 	appentitybase "github.com/openmeterio/openmeter/openmeter/app/entity/base"
 	appobserver "github.com/openmeterio/openmeter/openmeter/app/observer"
-	"github.com/openmeterio/openmeter/openmeter/appstripe"
-	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
+	appstripe "github.com/openmeterio/openmeter/openmeter/app/stripe"
+	appstripeentity "github.com/openmeterio/openmeter/openmeter/app/stripe/entity"
 	customerentity "github.com/openmeterio/openmeter/openmeter/customer/entity"
 )
