+++ openmeter/server/router/router.go
@@ -13,8 +13,8 @@ import (
 	"github.com/openmeterio/openmeter/api"
 	"github.com/openmeterio/openmeter/openmeter/app"
 	apphttpdriver "github.com/openmeterio/openmeter/openmeter/app/httpdriver"
-	"github.com/openmeterio/openmeter/openmeter/appstripe"
-	appstripehttpdriver "github.com/openmeterio/openmeter/openmeter/appstripe/httpdriver"
+	appstripe "github.com/openmeterio/openmeter/openmeter/app/stripe"
+	appstripehttpdriver "github.com/openmeterio/openmeter/openmeter/app/stripe/httpdriver"
 	"github.com/openmeterio/openmeter/openmeter/credit"
 	creditdriver "github.com/openmeterio/openmeter/openmeter/credit/driver"
 	"github.com/openmeterio/openmeter/openmeter/credit/grant"
